# docs/swagger.yaml
basePath: /api/v1
definitions:
  subscription.CreateRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  subscription.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  subscription.Response:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  subscription.TotalResponse:
    properties:
      total:
        type: integer
    type: object
  subscription.UpdateRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST service for managing user online subscriptions and calculating
    subscription totals.
  title: Subscriptions Aggregation API
  version: "1.0"
paths:
  /subscriptions:
    get:
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subscription.Response'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscription.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscription.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscription.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Calculates total subscription cost for an inclusive month period.
        Date format is MM-YYYY.
      parameters:
      - description: Period start in MM-YYYY format
        in: query
        name: from
        required: true
        type: string
      - description: Period end in MM-YYYY format
        in: query
        name: to
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.TotalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscription.ErrorResponse'
      summary: Calculate subscriptions total
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"